Create Queue
Studying algorithms and data structures I decided to write a fun queue class.  
A List object would hold the "history" of the input. 
The constructor initializes the list object //currentHistory.Add("History started: " + DateTime.Now); to indicate the beginning of the queue.
CollectHist() will add input to the list object.
PerLineHistory() and GetWholeHistory() provides two ways of recieving the history, PerLineHistory()  returns a line of input at a time while GetWholeHistory()  returns the entire history as an array.